% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatEnrichmentmap.R
\name{formatEnrichmentmap}
\alias{formatEnrichmentmap}
\title{Format the Enrichment map - move all red (NES +ve nodes) to the right and
all blue (NES -ve nodes) to the left}
\usage{
formatEnrichmentmap(network = "current",
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}

\item{suid}{of the network that you want to get the view for}
}
\value{
server response
}
\description{
Format the Enrichment map - move all red (NES +ve nodes) to the right and
all blue (NES -ve nodes) to the left
}
